Fix a validation bug and update functionapp
Content-type isn't sent in the initial subscription validation request.
Moving the check down fixes this, as that's the codepath for receiving
and event.

Updated dependencies and host.json file.

Updated tsconfig.json. This is really for a future change, but nice to
have the same tsconfig.json for all versions.